+++ ton-wallet-app/src/screens/index.ts
@@ -3,4 +3,5 @@
  */
 
 export * from './WelcomeScreen';
+export * from './CreateWalletScreen';
 export * from './MainScreen';